version2
исправление замечаний
