+++ booklib2_2/Readme.md
@@ -34,7 +34,7 @@ comadd 2 Иван "очень длинно и много французкого"
 
 вывести список комментарий по идентификатору книги
 пример:
-bookcomg 2
+bookcom 2
