--- booklib2_2/src/main/resources/application.properties
@@ -1,1 +1,0 @@
-
