+++ booklib2_2/src/main/resources/i18n/messages.properties
@@ -19,3 +19,7 @@ BOOK_NOT_FOUND=Book {0} is not found.
 genresuccessmod=Genre {0} was modified.
 BOOK_COMMENT_EMPTY=Comment text is empty.
 commentsuccessadd=Comment was added.
+bookname=Title: 
+bookauthor=Author:
+bookgenre=Genre:
+comments=User annotation:
