+++ booklib2_2/src/main/resources/i18n/messages_ru_RU.properties
@@ -19,3 +19,7 @@ BOOK_NOT_FOUND=Книга {0} не найдена.
 genresuccessmod=Жанр {0} изменен.
 BOOK_COMMENT_EMPTY=Комментарий к книге не указан.
 commentsuccessadd=Коментарий добавлен.
+bookname=Название: 
+bookauthor=Автор:
+bookgenre=Жанр:
+comments=Отзывы пользователей:
