+++ booklib2_2/src/test/java/ru/otus/spring/booklib/service/BookServiceImplTest.java
@@ -3,28 +3,23 @@ package ru.otus.spring.booklib.service;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
 import org.springframework.boot.test.context.SpringBootTest;
-import ru.otus.spring.booklib.dao.CommentRepositoryJpa;
 import ru.otus.spring.booklib.domain.Author;
 import ru.otus.spring.booklib.domain.Book;
 import ru.otus.spring.booklib.domain.BookView;
-import ru.otus.spring.booklib.domain.Comment;
 import ru.otus.spring.booklib.error.BookError;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 @SpringBootTest
-class LibraryServiceImplTest {
+class BookServiceImplTest {
 
     @Autowired
-    private LibraryService service;
+    private BookService service;
     @Autowired
     private AuthorService authorService;
 
@@ -147,35 +142,7 @@ class LibraryServiceImplTest {
 
     }
 
-    @Test
-    @DisplayName("Добавить книгу и комментарии к ней")
-    void createBookWithComment() {
-        Book book = null;
-        try {
-            book = service.createBook("тестовая книга c комментариями", "Новикович Автор Сочинитель", "Роман"
-                    , 0L, 0L);
-        } catch (BookError bookError) {
-            bookError.printStackTrace();
-        }
-        assertThat(book).isNotNull();
 
-        Long bookId = book.getId();
-
-        assertDoesNotThrow(() -> service.commentBook(bookId, "Вася", "Книга не понравилась, слишком много букв"));
-        assertDoesNotThrow(() -> service.commentBook(bookId, "Петя", "Очень интересная книга. Автор пиши еще."));
-
-        List<Comment> lstComment = new ArrayList<>();
-        try {
-            lstComment = service.getComment(bookId);
-
-        } catch (BookError bookError) {
-            bookError.printStackTrace();
-        }
-        assertThat(lstComment.size()).isEqualTo(2);
-
-        assertDoesNotThrow(() -> service.removeBook(bookId));
-
-    }
 
     @Test
     @DisplayName("Создать книгу, указав идентификатор автора")
